ADDS: i3-like window moving and focus
Can now move windows and focus across displays like in i3. Began attempting to script it in bash, but float conversion was complictaied, when I realized the yabai command will return a non-zero code if the window is alread at the targeted edge, so adding an or opperator will switch to targeting the display in the cardinal direction; move focus to left will attempt to do so in the display, or hit the or and move focus across the displays to the right most window on the left display.